DEEP-287: Rename bean to better reflect its responsibilities.
